requirements.txt
# PortWise AI - Python Dependencies
# Install with: pip install -r requirements.txt

# ── Core framework ────────────────────────────────────────────────────────────
streamlit==1.40.0
python-dotenv==1.0.1
pydantic==2.9.2

# ── LLM & Agent orchestration ─────────────────────────────────────────────────
langchain==0.3.7
langchain-google-genai==2.0.4
langgraph==0.2.45

# ── Document processing ───────────────────────────────────────────────────────
pdfplumber==0.11.4
Pillow>=10.0.0
pytesseract==0.3.13

# ── Voice / Audio ─────────────────────────────────────────────────────────────
SpeechRecognition==3.11.0
pydub==0.25.1
# Offline STT fallback — uncomment if you want whisper available without internet:
# openai-whisper==20231117

# ── ML / Analytics ────────────────────────────────────────────────────────────
numpy>=1.24.0
scikit-learn==1.5.2
pandas>=2.0.0

# ── Vector database ───────────────────────────────────────────────────────────
chromadb==0.5.18

# ── Visualisation ─────────────────────────────────────────────────────────────
plotly==5.24.1
altair==5.4.1

# ── HTTP ─────────────────────────────────────────────────────────────────────
requests==2.32.3
httpx==0.27.2
aiohttp==3.11.2

# ── Observability & logging ───────────────────────────────────────────────────
structlog==24.4.0

# ── Security ──────────────────────────────────────────────────────────────────
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.2.0

# ── Development / Testing ─────────────────────────────────────────────────────
pytest==8.3.3
pytest-mock==3.14.0

# NOTE: Do NOT pin 'asyncio' — it is part of the Python standard library
# and cannot be installed from PyPI. The old entry caused confusing install errors.
